refactor: move stuff around a bit
