+++ lib/NonlinearSolveBase/src/descent/geodesic_acceleration.jl
@@ -54,7 +54,7 @@ function InternalAPI.reinit_self!(cache::GeodesicAccelerationCache; p = cache.p,
     cache.last_step_accepted = false
 end
 
-NonlinearSolveBase.@internal_caches GeodesicAccelerationCache :descent_cache
+@internal_caches GeodesicAccelerationCache :descent_cache
 
 function get_velocity(cache::GeodesicAccelerationCache)
     return SciMLBase.get_du(cache.descent_cache, Val(1))
