+++ lib/NonlinearSolveBase/src/descent/newton.jl
@@ -24,7 +24,7 @@ supports_line_search(::NewtonDescent) = true
     normal_form <: Union{Val{false}, Val{true}}
 end
 
-NonlinearSolveBase.@internal_caches NewtonDescentCache :lincache
+@internal_caches NewtonDescentCache :lincache
 
 function InternalAPI.init(
         prob::AbstractNonlinearProblem, alg::NewtonDescent, J, fu, u; stats,
